🔧 chore(server.ts): change port variable case from lowercase port to uppercase PORT
The port variable in the server.ts file has been changed from lowercase "port" to uppercase "PORT" for consistency with naming conventions. This change improves code readability and maintainability.

🔧 chore(page.tsx): add TextWithImage component

The TextWithImage component has been added to the page.tsx file in the business/(results) directory. This component allows for displaying text content with an optional accompanying image. It enhances the layout and presentation of the page.

🔧 chore(page.tsx): add TextWithImage component

The TextWithImage component has been added to the page.tsx file in the business/(results) directory. This component allows for displaying text content with an optional accompanying image. It enhances the layout and presentation of the page.

🔧 chore(page.tsx): add TextWithImage component

The TextWithImage component has been added to the

🔀 chore(results-business.json): update textWithImage content and imagePath
The content and imagePath of the textWithImage section in the results-business.json file have been updated. The new content provides information about the benefits of pursuing an MBA or a Master's in Organizational Leadership, while the imagePath points to an image of a CEO. This update enhances the information provided in the application's business results section.