🐛 fix(server.ts): change port variable case from lowercase port to uppercase PORT ✨ feat(server.ts): add support for process.env.PORT environment variable The port variable is now named PORT, which improves consistency with the naming conventions as PORT is a constant. Support for an environment variable allows the application to be more flexible as it can now run on any available port specified via the process.env.PORT environment variable.
🐛 fix(route.ts): remove console.log statements
The console.log statements were added for debugging purposes and are no longer needed.

🐛 fix(page.tsx): set vertical value in useUser context
The setVertical function is now used to set the vertical value in the useUser context. This ensures that the correct vertical is used when making API requests.

🐛 fix(page.tsx): remove console.log statements
The console.log statements were added for debugging purposes and are no longer needed.

🐛 fix(page.tsx): remove console

🔇 chore(slugs.js): remove console.log statement
The console.log statement on line 10 has been removed to improve code cleanliness and remove unnecessary logging.