🔧 chore(next.config.js): add new hostnames to the allowed domains 🐛 fix(SchoolCarousel.tsx): remove unnecessary empty fragment The next.config.js file has been updated to include two new hostnames, "advance-cms.appily.com" and "**-eab-agency.vercel.app", in the allowed domains list. This allows requests to be made to these domains without any CORS issues.
In the SchoolCarousel.tsx file, the unnecessary empty fragment has been removed. This improves the code readability and removes any potential confusion caused by the empty fragment.

🐛 fix(getMatchedSchool.ts): fix formatting and add missing import
✨ feat(getMatchedSchool.ts): add support for matching states and fetching carousel cards

The code in `getMatchedSchool.ts` has been formatted to improve readability. Additionally, a missing import for the `Partner` type from `payload-types.ts` has been added.

The `getMatchedSchool` function now supports matching states against a predefined list of states and fetching carousel cards based on the matched states. If a state is provided, it will be matched against the `statesToMatchAgainst` object to determine the array of states to pass to `fetchCarouselCards`. If no state is provided, a default array of states is used.

If the fetched `cards` array has less than 5 elements, the function will also fetch additional carousel cards using `fetchReaminingCarouselCards` and append them to the existing cards. The combined array is then shuffled using the