🔀 chore(links-healthcare.json): update link objects in healthcare links data 🔀 chore(PageHeader.jsx): import NavBar from correct location 🔀 chore(Header.tsx): import NavBar from correct location
The link objects in the healthcare links data have been updated to use "href" and "label" instead of "slug" and "title" for consistency and clarity.

In the PageHeader component, the NavBar component is now imported from the correct location.

In the Header component, the NavBar component is also imported from the correct location.