🐛 fix(page.tsx): fix typo in high school resources link 🚀 feat(page.tsx): add link to homepage on main logo in header component 🐛 fix(Header.tsx): fix typo in Link type definition 🐛 fix(careers-business.json): fix typo in testimonial author title
In the `page.tsx` file, a typo in the high school resources link has been fixed to correctly point to the "appily.com" website.

In the `Header.tsx` file, a link to the homepage has been added to the main logo in the header component.

In the `Header.tsx` file, a typo in the Link type definition has been fixed.

In the `careers-business.json` file, a typo in the testimonial author title has been fixed.