🔧 fix(page.tsx): update title variable to "Build your plan" for better clarity 🔧 fix(page.tsx): update VERTICAL variable to "adcplan" for more accurate data fetching 🔧 fix(Quiz.js): uncomment import statement for Score component ✨ feat(Quiz.js): add logic to redirect to "/appily-redirect" if answer value is 1 for associatedField "age_range" 📝 chore: add new quiz data file quiz-adcplan.json The changes in `page.tsx` update the `title` variable to "Build your plan" to provide a more descriptive title for the page. The `VERTICAL` variable is also updated to "adcplan" to ensure the correct data is fetched for the quiz.
In `Quiz.js`, the import statement for the `Score` component is uncommented to enable its usage. Additionally, a new logic is added to redirect to "/appily-redirect" if the answer value is 1 for the associatedField "age_range".

The new file `quiz-adcplan.json` is added to the `data` directory, which contains the quiz questions and answers data for the "adcplan" vertical.